fixed bug with deleted packing slips
when deleting packing slips we instead set the field to "isPastVersion:
true" so that no "current" packing slips show up when searched, but we
still have a version to query.

having documents with this add-on field, has been presenting issues in
filtering and aggregating packed qties.

Here, when we delete a packing slip, b/c a document exists that matches
the items in it (when looking at packing queue) but it is also a past
version, it REMOVED the work order entirely from the queue because it
matched the filter incorrectly.

By removing the document from appearing in the primary pipeline in the
first place, the original filters work as expected.